0718 sql 함수들 models로 분리
